Refactored to new structure
Also added in new rule to handle and event target to handle legacy sync requests
