+++ infrastructure/stage/dynamodb/interfaces.ts
@@ -1,0 +1,6 @@
+import { RemovalPolicy } from 'aws-cdk-lib';
+
+export interface BuildTableProps {
+  tableName: string;
+  tableRemovalPolicy?: RemovalPolicy;
+}
