+++ infrastructure/stage/event-bus/interfaces.ts
@@ -1,0 +1,4 @@
+export interface BuildEventBusProps {
+  eventBusName: string;
+  eventBusDescription?: string;
+}
