+++ infrastructure/stage/lambda/interfaces.ts
@@ -1,0 +1,52 @@
+/* Lambda interfaces */
+import { PythonFunction } from '@aws-cdk/aws-lambda-python-alpha';
+
+export type LambdaNameList =
+  | 'findSinglePartFiles'
+  | 'generateCopyJobList'
+  | 'launchIcav2Copy'
+  | 'uploadSinglePartFile'
+  | 'checkJobStatus';
+
+/* Lambda names array */
+/* Bit of double handling, BUT types are not parsed to JS */
+export const lambdaNameList: Array<LambdaNameList> = [
+  'findSinglePartFiles',
+  'generateCopyJobList',
+  'launchIcav2Copy',
+  'uploadSinglePartFile',
+  'checkJobStatus',
+];
+
+/* We also throw in our custom application interfaces here too */
+export interface LambdaRequirementProps {
+  needsIcav2AccessToken: boolean;
+}
+
+export type LambdaToRequirementsMapType = { [key in LambdaNameList]: LambdaRequirementProps };
+
+export const lambdaToRequirementsMap: LambdaToRequirementsMapType = {
+  findSinglePartFiles: {
+    needsIcav2AccessToken: true,
+  },
+  generateCopyJobList: {
+    needsIcav2AccessToken: true,
+  },
+  launchIcav2Copy: {
+    needsIcav2AccessToken: true,
+  },
+  uploadSinglePartFile: {
+    needsIcav2AccessToken: true,
+  },
+  checkJobStatus: {
+    needsIcav2AccessToken: true,
+  },
+};
+
+export interface BuildLambdaProps {
+  lambdaName: LambdaNameList;
+}
+
+export interface LambdaObject extends BuildLambdaProps {
+  lambdaFunction: PythonFunction;
+}
