+++ infrastructure/stage/utils/index.ts
@@ -1,0 +1,4 @@
+/* Utils */
+export function camelCaseToSnakeCase(camelCase: string): string {
+  return camelCase.replace(/([A-Z])/g, '_$1').toLowerCase();
+}
