+++ package.json
@@ -41,7 +41,7 @@
     "@aws-cdk/aws-lambda-python-alpha": "2.195.0-alpha.0",
     "@aws-cdk/aws-pipes-alpha": "2.195.0-alpha.0",
     "@aws-cdk/aws-pipes-sources-alpha": "2.195.0-alpha.0",
-    "@orcabus/platform-cdk-constructs": "0.0.6",
+    "@orcabus/platform-cdk-constructs": "0.0.10-beta.1",
     "aws-cdk-lib": "^2.195.0",
     "constructs": "^10.4.2",
     "sqs-dlq-monitoring": "^1.2.21",
